--- a/app/src/main/java/com/gh4a/activities/PullRequestActivity.java
+++ b/app/src/main/java/com/gh4a/activities/PullRequestActivity.java
@@ -541,7 +541,7 @@ public class PullRequestActivity extends BaseFragmentPagerActivity implements
 
     private void loadPendingReview(boolean force) {
         String ownLogin = Gh4Application.get().getAuthLogin();
-        PullRequestReviewService service = ServiceFactory.get(PullRequestReviewService.class, force);
+        PullRequestReviewService service = ServiceFactory.get(PullRequestReviewService.class, force, ApiHelpers.MAX_PAGE_SIZE);
 
         ApiHelpers.PageIterator
                 .first(page -> service.getReviews(mRepoOwner, mRepoName, mPullRequestNumber, page),