--- a/app/src/main/java/com/gh4a/activities/RepositoryActivity.java
+++ b/app/src/main/java/com/gh4a/activities/RepositoryActivity.java
@@ -377,8 +377,9 @@ public class RepositoryActivity extends BaseFragmentPagerActivity implements
             showRefSelectionDialog();
         } else {
             final RepositoryBranchService branchService =
-                    ServiceFactory.get(RepositoryBranchService.class, false);
-            final RepositoryService repoService = ServiceFactory.get(RepositoryService.class, false);
+                    ServiceFactory.get(RepositoryBranchService.class, false, ApiHelpers.MAX_PAGE_SIZE);
+            final RepositoryService repoService =
+                    ServiceFactory.get(RepositoryService.class, false, ApiHelpers.MAX_PAGE_SIZE);
 
             Single<List<Branch>> branchSingle = ApiHelpers.PageIterator
                     .toSingle(page -> branchService.getBranches(mRepoOwner, mRepoName, page))