--- a/app/src/main/java/com/gh4a/adapter/CommitNoteAdapter.java
+++ b/app/src/main/java/com/gh4a/adapter/CommitNoteAdapter.java
@@ -204,7 +204,7 @@ public class CommitNoteAdapter extends RootAdapter<GitComment, CommitNoteAdapter
     @Override
     public Single<List<Reaction>> loadReactionDetails(ReactionBar.Item item, boolean bypassCache) {
         final GitComment comment = ((ViewHolder) item).mBoundItem;
-        final ReactionService service = ServiceFactory.get(ReactionService.class, bypassCache, ApiHelpers.MAX_PAGE_SIZE);
+        var service = ServiceFactory.getForFullPagedLists(ReactionService.class, bypassCache);
         return ApiHelpers.PageIterator
                 .toSingle(page -> service.getCommitCommentReactions(mRepoOwner, mRepoName, comment.id(), page));
     }