--- a/app/src/main/java/com/gh4a/fragment/CommitNoteFragment.java
+++ b/app/src/main/java/com/gh4a/fragment/CommitNoteFragment.java
@@ -233,7 +233,7 @@ public class CommitNoteFragment extends ListDataBaseFragment<GitComment> impleme
     @Override
     protected Single<List<GitComment>> onCreateDataSingle(boolean bypassCache) {
         final RepositoryCommentService service =
-                ServiceFactory.get(RepositoryCommentService.class, bypassCache);
+                ServiceFactory.get(RepositoryCommentService.class, bypassCache, ApiHelpers.MAX_PAGE_SIZE);
 
         return ApiHelpers.PageIterator
                 .toSingle(page -> service.getCommitComments(mRepoOwner, mRepoName, mObjectSha, page))