… app/src/main/java/com/gh4a/fragment/IssueFragmentBase.java
… app/src/main/java/com/gh4a/fragment/IssueMilestoneListFragment.java
… app/src/main/java/com/gh4a/fragment/PullRequestConversationFragment.java
… app/src/main/java/com/gh4a/fragment/PullRequestFilesFragment.java
… app/src/main/java/com/gh4a/fragment/RepositoryFragment.java
… app/src/main/java/com/gh4a/fragment/ReviewFragment.java
… app/src/main/java/com/gh4a/fragment/UserFragment.java
… app/src/main/java/com/gh4a/model/TimelineItem.java
… app/src/main/java/com/gh4a/resolver/CommitCommentLoadTask.java
… app/src/main/java/com/gh4a/resolver/CommitDiffLoadTask.java
… app/src/main/java/com/gh4a/resolver/PullRequestDiffCommentLoadTask.java
… app/src/main/java/com/gh4a/resolver/PullRequestDiffLoadTask.java
… app/src/main/java/com/gh4a/resolver/PullRequestReviewCommentLoadTask.java
… app/src/main/java/com/gh4a/resolver/PullRequestReviewDiffLoadTask.java
… app/src/main/java/com/gh4a/resolver/RefPathDisambiguationTask.java
… app/src/main/java/com/gh4a/utils/SingleFactory.java