@Fs00 2025-02-08
I have also added a commit to leverage `PageIterator.first` in the `RefPathDisambiguationTask` to avoid loading all branches/tags unnecessarily when resolving a blob/tree URL, which should help in situations like the one in #1252.